+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Next Release
 - Add minimum schedule value validator.
 Now it is possible to set minimum time frame for resource schedule operations.
 - Drop django-gm2m dependency.
+- Replace django.conf.urls.patterns with list of urls.
 
 Release 0.125.0
 ---------------